requirements.txt
###### Requirements without Version Specifiers ######
# --upgrade
pygooglenews 
newspaper3k
#nltk
nltk
textblob